fix(template): merge duplicate class attrs, multi-root html fails, schema dep docs
DX improvements:
- Duplicate class attributes merged with space separator (div(className := "a", className := "b") renders class="a b")
- html"" with multiple top-level nodes now throws IllegalArgumentException instead of silently wrapping in <span>
- Schema dependency documented in template.md
